Remove ArgumentConvert trait and implementations
